project/requirements.txt
streamlit>=1.37.0
google-genai>=0.3.0
python-dotenv>=1.0.1
scikit-learn>=1.4.2
numpy>=1.26.4
pandas>=2.2.2
PyYAML>=6.0.1
pytest>=8.3.0
ruff>=0.5.0
